refactor: simplify useFlag hook and introduce useFlagWithManualTracking for improved tracking control
